Do rounding explicitly.
git-svn-id: https://svn.dealii.org/trunk@22633 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/source/lac/lapack_full_matrix.cc
@@ -261,7 +261,7 @@ LAPACKFullMatrix<number>::compute_svd()
 	&work[0], &lwork, &ipiv[0], &info);
   Assert (info==0, LAPACKSupport::ExcErrorCode("gesdd", info));
 				   // Now resize work array and
-  lwork = work[0] + .5;
+  lwork = static_cast<int>(work[0] + .5);
 #else
   int lwork = 3*std::min(mm,nn) +
 	      std::max(std::max(mm,nn),4*std::min(mm,nn)*std::min(mm,nn)+4*std::min(mm,nn));
